refactor: move cheap unit level tests under lib
