+++ packages/@cdktf/hcl2cdk/lib/__tests__/coerceType.test.ts
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc
 // SPDX-License-Identifier: MPL-2.0
-import { coerceType } from "../lib/coerceType";
+import { coerceType } from "../coerceType";
 import template from "@babel/template";
 import generate from "@babel/generator";
